WIP Shows returning "shared" flag at connection create time
The default is to say it is shared, which means whoever is
using this needs to be more cautious with host actions. It
shows how something like LVMISCSI + LIO can override this
since they are unique, and how something like Pure drivers
can *always* specify it to be True (not really required,
more like documentation purposes on the driver).

One concern is where that gets stored... it looks like
we try to save it on the attachment object but where
does that end up? Do we just rely on the caller to
persist that info? Or should later GET's of the attachment
show that connection info?