ActiveCluster support for Pure Storage drivers
This adds sync-rep support for cheesecake api's via
Pure's ActiveCluster feature leaning on stretched Pods.

This is done by specifying the "type" of replication on
"replication_device" options in cinder.conf to be 'sync'.
With an optional 'uniform' boolean.

One thing to note is that we needed to start using
the provider_id field on volumes. With volumes in Purity
Pods the names change, so our quick formula for name
prediction goes out the window. Luckily there is already
a field on the volume objects for this.. provider_id.
Any volumes without one already set will get one updated
when the driver starts up.

Depends-On: https://review.openstack.org/#/c/571332/
Implements: blueprint pure-sync-rep
Change-Id: I135d2e5aa8a0a82ddd212eb8eba393e2fbef73ca